+++ android/src/main/java/com/pinmi/react/printer/adapter/PrinterAdapter.java
@@ -23,7 +23,7 @@ public interface PrinterAdapter {
 
     void printRawData(String rawBase64Data, Promise promise);
 
-    void printImageData(String imageUrl, Promise promise);
+    void printImageData(String imageUrl, double imageWidth, Promise promise);
 
-    void printQrCode(String qrCode, Promise promise);
+    void printQrCode(String qrCode, double qrSize, Promise promise);
 }